revert: service worker changes
